IDEA-266608 Restore time tracking in YouTrack integration, add plugin ads
Advertising for the YouTrack plugin is implemented as a message in settings,
and as a notification on submitting a work item. The notification balloon is
shown once per IDE session unless either manually disabled or the plugin is
already installed.

I've tried to implement a notification action to automatically install
the plugin, but, unfortunately, this platform API at the moment is quite
limited, so there are now just links to Marketplace.

IJ-MR-15066

GitOrigin-RevId: 2a77c66a3abb7ff48384ae86ded7ab08895dcc13