IDEA-266608 Restore time tracking in YouTrack integration, add plugin ads
Advertising for the YouTrack plugin is implemented as a message in settings,
and as a notification on submitting a work item. The notification balloon is
shown once per IDE session unless either manually disabled or the plugin is
already installed.

I've tried to implement a notification action to automatically install
the plugin, but, unfortunately, this platform API at the moment is quite
limited, so there are now just links to Marketplace.

GitOrigin-RevId: d17d1ea5f46e1627a30712ac5f8f37656834418c