EA-286009 Fixed run anything actions deferred execution
The reason for the bug is that our doWhenFocusSettlesDown may run Runnable synchronously or may put it in the queue.
In the first case everything works just fine, because we are under invokeLater and therefore write-safe context
In the second case, runnable is going to be processed on the last focus event, when queue is empty, and it will happen without invokeLater and without write-safe context.

I'm not sure if we should run all queued runnables in write-safe context in such case.

GitOrigin-RevId: 6e52667cb2d7825b2d49c05b4e75e06bc8a6f2b8