[markdown] convert owasp_java_html_sanitizer from project-level to module-level library
It's used from single module only so it makes no sense to declare it as a project-level library. Also, this will mean that its JAR will be included into the plugin distribution, not to IDE_HOME/lib directory. This way it won't be added to classpath of all other plugins.

GitOrigin-RevId: 0805d6aa8bc7270326aa0c8213856ad6003a002e