[vcs-log] hide text field popup on log tab disposal to avoid memory leak
SearchTextField creates an AbstractPopup instance in order to show search history, when setHistoryPropertyName method (called in constructor) triggers loading history and a fireContentsChanged call. This popup is supposed to be disposed in removeNotify method, but if the text field does not get shown it is not called. This can happen, for example, for additional Log tabs which are created automatically on start, and then could be disposed automatically on VCS root change. But as popup stays in memory and keeps a reference to the corresponding tab via ok handler, closed tabs occupy memory as well.

This commit adds a hack to manually hide the popup on Log tab disposal.

IDEA-273465

(cherry picked from commit 6ecc1c583d89ee653dc6fcdc51f44dd117a386b9)

GitOrigin-RevId: 070612b60976f0b8fe65f637a3f158ebd29bb820