[git-branches-dashboard] cleanup state management in branches tree
When branches tree is refreshed initially, its state is loaded from the BranchesTreeStateHolder, and if no state is saved there, the tree is just expanded. On other runs, tree state is saved into a variable and then applied at the end.

Before this commit, BranchesTreeStateHolder was used instead of the variable, but this did not work for all but the last opened tree, since BranchesTreeStateHolder only works with the tree stored in it.

GitOrigin-RevId: d8bf66722cc8aa6b1a5680140918d7973e92aece