[git-branches-dashboard] do not store branches tree in the BranchesTreeStateHolder service
Instead of keeping branches tree in the service and thus leaking the VcsLogUi instance it has reference to, the tree itself can persist its state when it changes. In case of multiple Log tabs and multiple trees this is also a bit more fair: the tree which is expanded or collapsed last gets its state saved, instead of the tree in the last opened tab.

IDEA-273330

GitOrigin-RevId: 29e5ec03b2c975bfce950d89c11a0173fb6681cb