IDEA-265812 add API to check whether a test should be navigated to immediately when it's the sole target
This reverts the "Go to Test" navigation behaviuor to what it was prior (never navigate immediately, always show the popup if there is a test creator). It's now possible however for individual test finders to customize whether they wish to navigate immediately, and they may select their own criteria to do so (such as checking the expected name, or simply whether the target is an actual test).

GitOrigin-RevId: eccf4d8e56adcb34f2dbcf5fb7dca0bf6ca31dad