RUBY-28197: navigation to related symbols should be immediate if there's just one symbol.
RBS navigation targets don't have PSI elements associated with them but the `navigate` method for them is properly overloaded; it's not necessarily to check that an item has an element associated with it to call `navigate` on it. For items that do not override `navigate`, no exceptions will happen because the default implementation of `navigate` checks whether `getElement()` returns null. Items that do override `navigate` should take care of element nullability themselves.

IJ-CR-11783

GitOrigin-RevId: c970a4a988df063b5ddeb2b2ffd8e25e04cacd11