IDEA-271528 Guard against bad MavenDomDependency
This is not a fix for the underlying issue, which is in the
Maven plugin, but rather a workaround to at least avoid issues while
that issue is still around. See IDEA-273727 for news on that issue.

(cherry picked from commit 02474a342c0b8fb18aace65a8b99791e5d22ddf9)

GitOrigin-RevId: 83385fe82c66c270dfe58d5b42acd1eedfd399da