[workspace model] CPP-24609 Handle cases when `.iml` at CLion changed, and we reload project entities
Without this fix we can't correctly handle the case where .iml file for CLion module
was changed but `DummyParentEntitySource` was not used at replace by source. This
leads us to the exception at rBs operation. To fix it we added this entity source
to the rBs filter and add changed entities listener at `WorkspaceModelOCRootsSynchronizer` to
avoid storing module with `DummyParentEntitySource` at the main store.

GitOrigin-RevId: e1b97fcee6064ccd0fc6994336f7ccaf32ef8d64