[java][switch resolve] IDEA-273929 Good code red - cannot resolve symbol in switch statement group
IDEA used to allow resolving if a statement or an expression that uses a pattern variable in a switch case label follows the switch case label immediately. This led to errors in resolving references to pattern variables when their enclosing statement or expression is not the immediate right sibling of the case label.

Now IDEA looks for a switch case label among all the left siblings of the currently analyzing scope and enables resolving if a case label is found, and it's the same switch case label that is being analyzed.

(cherry picked from commit 758c34fde30db2fe9692d8076c49af2b3efae38f)

IJ-CR-12428

GitOrigin-RevId: aa92462b54fde6598ebcece36089df864938649e