[java][highlight] IDEA-274531 errors shown incorrectly when a method name is the same as a nested class name
IDEA added a fix to remove the "new" keyword when a callsite looks like a call to a method or a field. It caused the mentioned regression, because sometimes the callsite might simply instantiate a class that has the same name as a method which was mistakenly highlighted as an error. This patch simply adds a check if the resolved element is of a PsiClass and if so then the mentioned highlighting doesn't get added to such callsites.

(cherry picked from commit a34e6e1715d93ac21a2ba93cf5f57c54436bc11b)

IJ-CR-12429

GitOrigin-RevId: 5a3ab19185ccfdee8805adcc4116a977ec7bb27b