[java][switch] IDEA-273959 Good code red: 'cannot resolve symbol' in preceding case label
The resolver for PsiSwitchLabelStatement didn't use to consider the situation when there is "case null" after a case label with a pattern variable that is used in the "case null" clause. This patch adds check if the found switch case label is one of the special ones: "case null", "case default", "default" and if so it skips them when looking for a switch case label to resolve a reference against.

(cherry picked from commit 9f00bc64022ede4465ebf6e24f6165329f05a3e7)

IJ-CR-12432

GitOrigin-RevId: 7461968a7c971a17e2136ed47cf273683f46acd1