[workspace model] add methods to allow adding many dependencies to a module efficiently (IDEA-274168)
'addLibraryEntry' or 'addModuleOrderEntry' create a new copy of ModuleEntity in their implementation because currently the workspace storage doesn't allow modifying properties of entities in place. So these methods are slow if there are many dependencies in the module. The new 'addLibraryEntries' and 'addModuleEntries' methods will work much faster in such cases.

GitOrigin-RevId: 3c57360f9e54190d35d5fb3f39df5eb7868644c0