FIR/UAST: clarify parent/source uses in base service APIs
UAST element has a parent linkage if given, and for APIs that convert
the given PsiElement and/or retrieve a sub element, the given UElement
will be uased as a `parent`.

In contrast, UElement given for type conversion is usually an owner of
the PsiElement, and thus it's not quite `parent`. It should rather be
named `source`.

GitOrigin-RevId: f75745101f745fcb4d0ebe76d133104e75a1de1d