PY-49559 Strip trailing backslashes from the path to skeletons directory
It causes some weird problems with escaping, when in some places the platform
escapes `C:\foo bar\` as `"C:\foo bar\\"` and as `"C:\foo bar"\` in others
(compare ParametersListUtil.escape and CommandLineUtil.backslashEscapeQuotes).
One way or another, Python interpreter gets the path to the skeletons directory
as two separate command line arguments. The platform escaping problem needs
further investigation, but removing the trailing backslashes is the simplest
workaround on the Python plugin side. We stripped them in generator3 anyway.

GitOrigin-RevId: 52c0c1d3733cc004c1f0c2f1bf8cc186dfa698de