CWM-3374: "Client tries to hide host's lookup" fix.
Reason: there was an "another editor was selected" listener, which was triggered both on host's events and on client's events, but every time tried to close all host's hints (including lookups from the ticket).

Now there is the same listener, but on host's events it tries to close all host's hints, on client's events - client's hints.

GitOrigin-RevId: 2687cc12011e512de07a38d91cbb82a81f33b894