settings: add ability to change roaming type from PER_OS to DEFAULT without changing the xml name
If a component's Storage has the PER_OS roaming type, then it means that the data is saved in the `os` subfolder (e.g. options/mac/keymap.xml).

If we decide that the roaming type is incorrect, and the component should be shared between OSs or should not be shared at all, then we change the RoamingType either to DEFAULT or to DISABLED. We can't just change the RoamingType, because that would make the IDE forget the previous location. Instead, to migrate properly, we should mark the existing PER_OS storage as deprecated, and declare a new non-deprecated one with the correct RoamingType.

When the ComponentStore is saving components, it gets 3 storages for such a component: the new storage with the correct RoamingType, and two storages for the previous PER_OS roaming type: the actual 'os/name.xml', and the declared and automatically deprecated 'name.xml'.

Previously, we had to change the name of the xml together with the roaming type, because otherwise the ComponentStore would at first (1) save data to the 'name.xml' according to the new storage specification, then (2) remove the deprecated storage 'os/name.xml' (which is fine), and then (3) remove the double-deprecated storage 'name.xml', which is not fine because it is not deprecated anymore.

This fix prevents removing the storage xml, if it is already being used by a non-deprecated Storage. It allows to change the PER_OS roaming type without changing the name of the xml file.

GitOrigin-RevId: 3a8603f015f9d0c1089520bebe8d4c93a64b044a