[configuration store] get rid of noisy warnings in logs in some tests
"Content equals, but it must be handled not on this level" warning was shown in some tests which forces saving to check real format of saved files (e.g. IntelliJProjectFormatTest). It was quite noisy and complicates reading the test output. Now the message is printed with debug level in tests, so it'll be shown in TeamCity if the test fails, but won't pollute output otherwise.

GitOrigin-RevId: bcbcf24c85e5c5a3fe23e3fe1e66d4e0ca7017b4