[IntelliLang][injection] IDEA-276146 Slow operations are prohibited on EDT when injecting JAVA Language
During JAVA injection IDEA used to look for `annotations.jar` in order to see if it can modify the code to add the `@Language` annotation to the code. The search traverses all the jars in the project and hence it accesses the index data, which lead to the `Slow Operations are prohibited` error in the logs.

The patch moves to the background the search with the help from ProgressManager#run, which removes the index accessing logic from EDT

GitOrigin-RevId: ad35f2ba180028a504e9d030f909f4e29e103ecc