IDEA-275480 unconditionally turn off all grouping rules except for file grouping rule in `ShowUsagesSettings`
The bug was caused by 86eb345cb97f85f822169d112608656c93aba466. The groups were expected to be not even created, instead
they were instantiated but not rendered. That commit started rendering all groups. They cannot be turned on or off in
the UI, but the settings (restored from broken `usageView.xml`) turn them on causing the behaviour described in the
issue. This commit turns off the grouping, so that the groups are not even instantiated.

GitOrigin-RevId: b19d7fd615491d58147fa8e0c2777c97ae06aad2