[git-stash] read index and untracked changes in the stash
Stash commits are merge commits. First parent is the head of the branch, second -- commit containing index changes. If untracked files are saved in the stash, they are kept in the third commit.

In this commit parent commit hashes, except for the first one, are saved to the corresponding StashInfo and changes for them are cached in the GitStashCache.

GitOrigin-RevId: 961e5825619c3e4ef1062d0360c40ea4fca84efe