[git-stash] show correct notification when unstash is impossible due to index conflicts
When "stash pop" or "stash apply" command are called with the "--index" parameter, existing index content may prevent from applying stashed changes correctly. In this case conflicting files are not marked as unmerged and the conflict can't be resolved in a normal way, so this situation requires special treatment.

IDEA-276444

GitOrigin-RevId: e12c17d3369f8214f1ba8d1f9e3ce48022748eae