[Workspace Model] [IDEA-276464] Artifact properties were not loaded in workspace model
Previously due to the incorrect implementation, default artifact properties were not created and loaded for the artifacts.
In the implementation, on creation of the new artifact, or loading it from xml files, every artifact gets a list of default properties. These properties have no information and are not stored in artifact xml files.
However, this affects presentation, so if some property missing, the user won't be able to add or modify it (see the bug attached).

Another issue with properties processing was storing the default "unchanged" properties. Such properties should not be serialized on disk (xml files), they have a null value in PropertiesEntity, however, they still return a valid property when accessing from legacy bridge.

GitOrigin-RevId: 525e7d242e4ae25ba01da425e78f714ab12cd429