[yaml] return `YAMLKeyValue` declaration by `YAMLKeyValue` element
The declaration search walks up the tree from a leaf element up to the file,
and if a target is found by an element at any level, then the search stops.
`YAMLScalarKeyDeclarationSearcher` returns a target (`YAMLKeyValue`) by a leaf element (key) stopping the walk-up,
and effectively preventing any other targets to be provided by a parent element (`YAMLKeyValue`).

`YAMLKeyValue` has to implement `PsiDeclaredTarget` so that the platform doesn't treat the whole element range as a declaration.

GitOrigin-RevId: fb39f9fdea380d78768bd478b17c1fc6232ace22