settings-repository: remove dead code
There was no notification with the title of "VCS Update Finished" for a long time, so in fact autosyncing on this event wasn't happening, and we've received no complaints about it.

Autosync on push was happening, because the "Push successful" notification is being shown, but it seems to be inconsistent and confusing, so it is safe to remove that event as well.

GitOrigin-RevId: 4c3b89f3d4f9acca7c4f11fb363a77f93052a8a9