[vcs-log] navigate to the correct change from the "Go to Changed File..." popup
In merge commits a file could be changed to several commit parents, in which case there would be several changes for it in the browser. In order to correctly identify, which change to select, this commit first finds a node corresponding to the specific parent using a provided Tag object, then searches for the change among the children of that node.

GitOrigin-RevId: 56aece78dfa9e42739198710ce766e52a6092c68