[PSI Viewer] IDEA-277101 Slow operations are prohibited when selecting a word in PSI Viewer
Slow Operations are prohibited exceptions in PSI Viewer used to happen when a IDEA was trying to find an element at the offset or when it was looking for a common parent for injected languages. Such problems are elimited by moving these operations to a background thread with the help from ProgressManager#run.

GitOrigin-RevId: d33ebc443b81973eed1fc658f3a56f10451a4cd1