[groovy] IDEA-275688: Avoid compilation issues
Somehow GrPermitsClauseElementType refused to compile due to unimplemented abstract method, that was obviously implemented in one of classes in the hierarchy. I suspect that it is a bug in JPS, and I'd like to return to it one day. Nevertheless, this strange trick I did now is sufficient to resolve the problem.

GitOrigin-RevId: 9ec2313d3fbbedb9274029fe7c4f169e354439b0