[project model] don't fire rootsChanged event when changes in project configuration files are detected (IDEA-277398)
When a project is saved for the first time, its iml files are written via VFS, this may cause VFS events to be fired, and their paths may be references from entities via 'entitySource' properties. We should ignore such changed to avoid unnecessary 'rootsChanged' events which may cause performance problems if many iml files are generated.

If an iml (or other project configuration file) is changed externally, JpsProjectModelSynchronizer will detect that and update the workspace model, so RootsChangeWatcher will be able to handle real changes in the project configuration (if any) anyway.

GitOrigin-RevId: 60c24597af8892662bbc04c0557ac218652a500f