[file-history] correctly show diff for commits where file was added
If file history is requested when log is not ready, data pack is built from the data collected from calling git. But since this data does not contain information about connections between commits in file history, none of the nodes in the resulting graph have any parents. To work around this limitation, when a commit in file history has no parents, diff is shown with the previous commit in the table. However, in normal file history graphs there can be commits that do not have parents and are also not the first in the table. This can happen when the same file was added in two different branches. Comparing such commits with the previous commit in the table is incorrect.

To distinguish between these two cases, this commit adds "NO_PARENTS_INFO" key to mark visible packs without parents information.

GitOrigin-RevId: c0456dc4502796b6d0eaa76d80211c4ef695f955