ML in SE: Remove dependency on intellij.java.testFramework (IDEA-274480)
Replace with the dependency on intellij.platform.testFramework,
and change the SearchEverywhereFileFeaturesProviderTest accordingly.

Because PsiUtil was able to set the file type based on the extension,
the code has to be changed, as file type has to be explicitly specified,
otherwise files cannot be opened in the editor as their file type is unknown.

To avoid adding any unnecessary dependencies just for file types,
the main file type used is now a plain text file.

GitOrigin-RevId: ec91aaf7288ad9b453599f9d9aa188309b55f823