[groovy] IDEA-275560: Fallback to forked block element types for switch expressions
The previous approach to the problem of memorizing the "inside switch expression" key consisted of forbidding skipping lazy-parseable blocks within switch expression. It was not successful since reparse is bound to lazy-parseable blocks, and if it was invoked then the block was reparsed as if it is a usual not-inside-switch one.
Seems like it is no way to solve this problem now (without touching the platform) other than forking necessary block element types.

GitOrigin-RevId: dd88e05005060a2af6d4d87336325281f6c5a66c