[file-history] show commits collected from history provider as soon as they appear
Instead of waiting for full file history to show it, run file history task in another thread. Wait for a small timeout for new commits and as soon as they appear build a visible graph for them to display, while the task still runs. The task is stored in FileHistoryFilterer to allow getting commits from it when more is requested.

GitOrigin-RevId: 2ff247eb86d4c280647ccc7b8f7486f0b1552763