[git-index] correctly select next file from the diff when file status changes
When changes in the diff are added to the index, file can change its status. Previously it prevented "Next Difference" action from finding current GitFileStatusNodeWrapper instance in the list, as status was used for comparison. This commit changes equals and hash code for GitFileStatusNodeWrapper so that only node kind and file path are compared.

GitOrigin-RevId: aafddcb2efbaed26364947a04ea53eef71b170de