CWM-2147 optimize JComponents storage in HighlightInfo (review fixes IDEA-CR-71932)
In most cases there is only 1 FileEditor per file, because several FileEditor's per file can be created only when Code With Me session is launched and several clients open the same file. It was not good to use the Map as a storage with single key-value pair in most cases. Now the Map is used only if 2 or more key-value pairs exist.

GitOrigin-RevId: 9e66ea29df444f139be3e7b55c05ebda2b745763