IDEA-276966 don't throw `CancellationException` from `JobRunnable`
It's supposed to be the outer-most runnable, its thrown exceptions will go through
`AppScheduledExecutorService$BackendThreadPoolExecutor#afterExecute` and Thread's uncaught exception handler,
which is not desirable.
It's ok to throw from a callable returned from `JobFutureTask#jobCallable`, because the
exception will be caught by `FutureTask#run` and saved to be thrown from `Future#get`.

GitOrigin-RevId: 61d4d71dc67684ca65a3a3cbbb1e7a1936b68a0b