[slow operations] IDEA-277758 Slow operations are not reported in dialogs
Add the MODAL_ACTION_PERFORM activity in to check if it's allowed to perform slow operations under modal dialogs. The SlowOperations#assertSlowOperationsAreAllowed method filters out all the action.perform activities from the stack when the current modality is not NON_MODAL. It's safe, because when an action is performed and a modal dialog is shown this activity is recorded as modal.action.perform.

GitOrigin-RevId: 10e2f3fac21e649479f4914ddb34eb6ff84314f6