[completion] Fix LookupTest flakiness
It looks that even with delay 0 dispatchAllInvocationEventsInIdeEventQueue() does not guarantee that the alarm action will be executed. This was especially visible on Windows. Probably due to non-monotonic nanoTime() or higher granularity or something else?
Now, we just schedule the computation to EDT without any delay
Also, dispose lookup, just for the case.

GitOrigin-RevId: 5bf88e8087218a9a8733965a07982d3b1a1ec331