[java-analysis-api] Check for PCE before running visitor
The `PsiElementVisitor#visitElement` checks if the progress is cancelled, and all methods of the class are delegating its calls to it by default. The `AbstractBaseJavaLocalInspectionTool.buildVisitor` method defines a visitor that doesn't check if the progress is canceled which can cause freezes.

This patch adds calls to super methods of the visitor to implicitly check for PCE.

GitOrigin-RevId: ca718c4660f8f80adc5f9775dc6e1930a98b283f