HighLevelParamInfoHandler: Remove use of symbol pointers between findElementForParameterInfo() and updateParameterInfo().
It was unnecessary because candidates can't change for the lifetime of
the param info popup, and we resolve the call again anyway in
updateParameterInfo().

This also has the benefit of avoiding current issues in symbol
restoration (e.g., substitutions, static scope) and therefore we can
proceed with enabling other cases.