FIR IDE: register quickfix to specify type for AMBIGUOUS_ANONYMOUS_TYPE_INFERRED
FE1.0 also register the quickfix for
PROPERTY_WITH_NO_TYPE_NO_INITIALIZER and MUST_BE_INITIALIZED but it
doesn't make sense because there is no way to infer the type if there is
no information provided. The feature does not work in FE1.0 either.
Hence we are skipping those in FIR.

GitOrigin-RevId: 0524636add41fe41f8b2fb5eb8f6521be1aa940e