[vcs-log] simplify CommitDetailsGetter
SequentialLimitedLifoExecutor is useful for loading VcsCommitMetadata in MiniDetailsGetter when scrolling the table, but is not helpful in CommitDetailsGetter for loading VcsFullCommitDetails which are used in the change browser. It makes more sense to load full details only when they are explicitly requested by calling loadCommitsData. So in this commit CommitDetailsGetter#getCommitData does not start a loading task anymore, and either returns details from the cache or a placeholder commit (LoadingDetails instances, now kept in the separate cache). Without the need to track loading tasks for commits, there is no need restrict thread access to EDT.

GitOrigin-RevId: 79a637a8dcb0cb0c0ed433b036df632e41bd8616