EA-385744 - NPE: IdeKeyEventDispatcher.doUpdateActionsInner
`ActionUpdater.updateActionReal` returns null in case of exception, and that makes `session.presentation(action)` return the fresh copy of the default presentation which is not linked to an event. The logic seems reasonable for toolbars and menus but not for the shortcut processing.

(cherry picked from commit 3a3e6187a9ebd95e636000930056c6d4330bf3a9)

IJ-CR-15157

GitOrigin-RevId: be66f8523e8edf3bf26ba6394e098796959b25ff