IDEA-277609 [documentation] queue popup showing in the correct `CoroutineScope`
This fixes the `java.awt.IllegalComponentStateException` when the popup show event
was processed after the lookup was already closed.
After the change: closing the lookup cancels `showDocJob` in `autoShowDocumentationOnItemChange`,
which in turn cancels the scope in `handleAutoShowRequest`, which is now used to queue the popup showing.
In all other cases the popup showing is queued in the `DocumentationManager` scope.

GitOrigin-RevId: f84d73b1e0e67ab5cc71d5a3597691c5b26d2376