Add alternative lookup path for projector libs
In some IDEs (e.g. Android Studio) cwm-plugin-projector is not bundled. There should be a possibility to install projector libraries from non-bundled cwm-plugin.

cwm-plugin-projector layout is changed: all the projector libs are now placed to "lib/projector" (instead of "lib"). This is done to ease libraries lookup and filtering. Instead of assuming that we need all the libraries except those ending with "projector.jar" (which is simply not true if the projector libs were served from cwm-plugin/lib folder), we store all the projector libraries in a dedicated folder.

IJ-CR-15618

GitOrigin-RevId: b09c8a5025ca195ca7dcc49a56963b97f5e053a3