ML in SE: Add ranking-model for files. Remove lastOpenToolWindow. (IDEA-280849)
The lastOpenToolWindow has to be removed, as processing files (calculating weights)
takes place in a read action, which does not allow invokeAndWait to be used, which
is required for obtaining the lastOpenToolWindow, thus the feature is removed for now.

Additionally, the commit merges experimental and standard model providers for actions,
because the logic did not work in the case where there is just one (experimental) model.

(cherry picked from commit 54a2e2549fff896982ced429aed6b2140d84b497)

IJ-CR-15622

GitOrigin-RevId: af3b688d8a7b501f33813442db95e09b9a72400e