ML in SE: Add ranking-model for files. Remove lastOpenToolWindow. (IDEA-280849)
The lastOpenToolWindow has to be removed, as processing files (calculating weights)
takes place in a read action, which does not allow invokeAndWait to be used, which
is required for obtaining the lastOpenToolWindow, thus the feature is removed for now.

Additionally, the commit merges experimental and standard model providers for actions,
because the logic did not work in the case where there is just one (experimental) model.

GitOrigin-RevId: 3a67a91024a19f32ef0eef8156caa45d87d0f929