[build scripts] don't include a library to IDE_HOME/lib if it's added to a custom directory in plugin layouts
Strictly speaking, if a library is included into a subdirectory of 'lib' directory in a plugin layout, it won't be added to the classpath of the plugin's classloader, so we may need to put it into IDE_HOME/lib to include into the platform classloader. However, if a plugin explicitly includes a library into its layout in a non-standard way, most probably it means that it'll access it via dynamically configured classloader, so there is no need to include it to IDE_HOME/lib.

(cherry picked from commit 12722b79d99a8881ce3f91bd8094845efd28bbf6)

IJ-CR-15814

GitOrigin-RevId: 1099d8b17d66af8520bce9b78fbea94e9ad37cf9