[groovy] IDEA-221054: Process cycles in dfa as early as possible
Since control flow is in fact a graph with cycles, there should be determined some order of its processing. This order is generated by topological sorting of the control flow graph.
Consider the following control flow graph, where vertices are numbers:
1 -> 2; 1 -> 4; 2 -> 3; 3 -> 1; 4 -> 5.
Earlier, the generated order for forward DFA was [1, 4, 5, 2, 3] (2 and 3 were the first vertices the topological order was stopped on, and then the whole array for order was reversed). Notice, that the cycle (1, 2, 3) was processed the last. In actual control graph, this cycle could contain some instructions that would affect instructions with greater numbers, and then these later instructions had to be reprocessed once more.
This change moves cycles to the beginning of the order, i.e. in aforementioned example it would be [1, 2, 3, 4, 5], and therefore instructions 4 and 5 will be processed only once.

(cherry picked from commit 2f2906e13e266fa5896290083105bbe5a158af0e)

IJ-CR-16534

GitOrigin-RevId: a86f57a9630baa29d81f4f5830bb4497eabbe098