FIR Completion: Add `WithSuperDisambiguation` insertion strategy
This strategy is used if receiver is `super` and the inserted element
needs disambguation on the receiver type. For example,

public interface A {
    public fun foo(i: Int) {}
}

public interface B {
    public fun foo(i: Int) {}
}

class C : A, B {
    override fun foo(j: Int) {
        super.<caret>
    }
}

`super.foo` would be a compilation error because it's ambiguous. This
change allows the IDE to insert `super<A>.foo` or `super<B>.foo`
instead.

Note that the next commit actually make the change to use this
insertion strategy

GitOrigin-RevId: 8ec203b0e544ab343d48d7038447e6afdda972a4