[java-psi] TypeInfo: avoid hashCode calculation for most of the types
During indexing, type strings are mostly freshly-created strings, so we need to calculate hash-code for each of them to intern. We can pre-filter them by length before querying the hash-map to avoid hashCode calculation for most of the types
Only few lengths (3, 4, 5, 6, 7, and 16) are actually used for interning, so most class FQNs should be skipped now.

GitOrigin-RevId: 537a1f8808574e97f4132dd5071a4b2ffd1a806e