[kotlin] fix function breakpoints
The breakpoints are reloaded in the debugger thread under read action when the debugger attaches to a VM. It makes sense to not update the breakpoint name on EDT, since we should wait for the update to be done, or the breakpoint request won't be sent (since the name is null before the update). Changing the name in invokeAndWait may lead to possible deadlocks. MethodBreakpoint has the same implementation of name update.

^KTIJ-20001 Fixed

(cherry picked from commit 49567210e27bc2144d37ec0741683b6c47c98cdb)

IJ-MR-16750

GitOrigin-RevId: 328ae8e48edf4c6c76077ae8ce58160d2fdebced