trusted-projects: always display an editor notification in the safe mode
Previously the notification was displayed by each subsystem individually (namely, external systems, maven, ruby). Since IDEA-278339 it makes sense to always show the notification to allow the user to move from the safe mode to the trusted mode.

Clicking to the "Trust" button in the notification makes the project trusted, which generates the trust state change event.

However, in some cases (e.g. for external systems) we need to know the origin of the project trust event: if the project became trusted because of some action covered by a "do you trust this project" confirmation, or from this editor notification button. For such cases a separate method in the TrustChangeListener is introduced, with a bunch of helper methods.

(cherry picked from commit 79d03a5d7436d3e3bf6fd0f35cfbf49a4448647f)

GitOrigin-RevId: 3e6c9c21095b2f31f7077f539f00272a3a4497ea