[vcs-log] do not start new tasks in containing branches listener
ContainingBranchesGetter does not guarantee that all submitted requests are going to be fulfilled, moreover it drops all but last 10 of them. Also, it notifies listeners about new branches after each processed commit. So when VcsLogCommitSelectionListenerForDetails requests the branches for the new selection and receives the notification, it's more than likely that not every commit in the selection will have branches computed. If this listener sends more requests for these commits, more notifications will be produced, causing the process to be repeated, consuming more cpu. Any action that changes branches state, such as checkout, will worsen the situation, since cache will get cleared, causing more requests. In order to avoid this problem, branches should be retrieved from the cache in the listener (as was done prior to the dd1d41a0ee91639b3908328c4d95bd2b3d3464e7).

IDEA-282260

GitOrigin-RevId: a34f0b90e1c11514215cd8f9658bf4064962d817