IDEA-276596 [documentation] update reusable tab (if any) before showing a popup
This change allows the following use case:
1. open popup, then move it to TW
2. turn off auto-update
3. move caret and invoke the action again
Before the change the popup was shown, after the change the shown reusable tab gets updated.

This changes makes the TW tab interaction similar to the previous quick-doc implementation for lookup, SE, and
rendered doc links: before the change the new popup was opened, after the change the shown reusable tab gets updated.

GitOrigin-RevId: fb05d36a73029cebc94d5c3428dd318b7a89ad8f