[IDEA-282620] Search popup starts search on each typed character
A speed search popup used to execute the search among the shown elements when a user typed a character, which led to small freezes when the list of elements is large. This patch adds two timers to delay the start of the search so the user can type the query before the search actually starts.

GitOrigin-RevId: 183f49b351e3aea0a930d1c0c3b9309fa185c6d9