[groovy] IDEA-273785: Remove shared variables handling
Shared variables (reassigned locals under @CompileStatic) are handled in two typecheck passes in the Groovy compiler. It is a big complication for our data flow, and so far there was no complains that inference should work differently. So I remove this code to simplify DFA and make it more maintainable

GitOrigin-RevId: 17d83cdccd676f3fb9dbf7d50fa1d861aa6eaeb0