[fleet] Fixes "Fleet doesn't recognize Rust project in remote environment"
The LSP backend was failing because the network call to check for the latest
version of Rust Analyzer was timing out some times (after two seconds), the
error was not being properly reported, which caused the LSP backend to be
red without an apparent reason.

This change makes sure that this kind of errors are properly reported, and
also that they don't cause other unexpected issues. Also the timeout to
check the version has been increased.

Fixes FLEET-T-4943

GitOrigin-RevId: 6eeca08000a4f72d3ce3d8743b1aad3163e229d0