Revert "git: execute clone with null project"
This reverts commit de63c2294355db102e31e0824d208f4567976a6c.
There is no need to support cloning from inside a project opened in the safe mode anymore, because all Git menus, actions and toolwindows are hidden from there.

At the same time the commit was breaking Git authentication during clone, because the code was expecting NotNull projects only. Fixing that is possible but requires a lot of changes and breaking the API.

GitOrigin-RevId: 262b1b89af559ee1a0062e5be4cddda51bfad894