[build] allow to provide environment to execute a project task FLEET-T-4855
Often `ProjectTask` can be run by `ProjectTaskRunner` using platform execution api.
This change provides an access to `ExecutionEnvironment` if it's supported by a `ProjectTaskRunner` w/o the `ProjectTask` execution. So, it can be reused for other use-cases. E.g. get the task run parameters and run it using another tool or environment.

GitOrigin-RevId: c71618f21b73b52cba177ded65774dbd432323fa