CPP-25966: Reformat completion snippets in more cases
We used to only indent compeltion results when we wanted
to keep the cursor position.

With the new custom reformat call in clangd we can reformat
the range while keeping the cursor position.

With the builtin formatter we instead split the range into lines and
reformat line by line starting from the end. The line with cursor
is only indented in this case.

GitOrigin-RevId: 52c75649b45b17de49fa930e48826aea42047d57