Replace 'firstNotNullResult' with 'firstNotNullOfOrNull'
S.Erokhin: I've roll-back and exclude most of changes from this commit during kt-213-master creation.
It was done because they were made in android plugin and that is outside the kotlin plugin.
Such changes are not allowed in the kt-* branches and it should be fixed by restoring the API in
the Kotlin compiler. It will be done later on

GitOrigin-RevId: 0d8e97eb8b27b51d79380054bbb8989c2b77e614