[fleet] Fixes IJ backend not loading files without document during rename.
When renaming something in fleet and some usages are in files not currently
loaded (i.e. without a DocumentEntity in the db), those files will be updated
but they will not be reported properly to the IJ backend, so the backend
thinks that the files remain unchanged and it will show zero usages for the
newly renamed thing.

This problem was caused by different issues:

First `FleetDocumentSyncVetoer.mayReloadFileContent` was always returning
false, which caused the IJ backend to never reload any file changed
outside of Fleet (or not in the Fleet db).

But changing `FleetDocumentSyncVetoer.mayReloadFileContent` to return
true was not enough. Because the way refactorings via IJ backend works
in Fleet is that we apply the refactoring, get a diff to construct an
`IdeaDiff` object, and then apply the diff in reverse to leave things
as they were before. Then Fleet applies the diff on disk, and we hope
that the `FSDFileWatcher` picks the change and reports back to IJ.

However, in the process of making the initial refactoring and applying
the diff in reverse, IJ's file manager marks the files as "unsaved".
When the changes are picked up by `FSDFileWatcher` and reported back,
the file manager will refuse to rescan the file because it's marked
as "unsaved".

To fix this we make sure that the files are not marked as unsaved
applyng the diff in reverse. For that we had to make
`FileDocumentManagerImpl.removeFromUnsaved` public (and marked as
internal).

fixes FLEET-T-4641

GitOrigin-RevId: 62894d6d2e4c78477ba0bd2595a4e32374391d4d