KTIJ-20144 MPP: "Nothing here" on test gutter click
The issue happened when a project had no jvm targets.

When a test class is opened in editor or the user clicks
on the run gutter button we try to find all gradle test
run tasks. Should a program element (class/method) has ones,
configuration(s) is(are) created, otherwise "Nothing here"
is displayed. [1] is a method responsible for listing the tasks
mentioned above. Kotlin-plugin implementations both for class
and method [2][3] didn't override it, hence leaving this
responsibility to its super class residing in Gradle-plugin.
The latter is unaware of targets other than jvm.

Now that [2] and [3] provide all tasks correctly another
problem arises. Common modules usually have multiple
alternatives and [4] was the only option to find them. Calls
to [4] lead to user interaction, which is unacceptable for an
editor opening (file analysis). Here is why [5] was introduced.

-------------------------------------------------------------------
[1]: AbstractGradleTestRunConfigurationProducer.getAllTestsTaskToRun
[2]: AbstractKotlinMultiplatformTestClassGradleConfigurationProducer
[3]: AbstractKotlinMultiplatformTestMethodGradleConfigurationProducer
[4]: MultiplatformTestTasksChooser.multiplatformChooseTasks
[5]: MultiplatformTestTasksChooser.listAvailableTasks

^KTIJ-20144 fixed

GitOrigin-RevId: 07acdda4d67aca354e33608b90298f268cdf3075