[speed search] EA-422029 NSEE: SpeedSearchBase.findPreviousElement
Use the new element iterator for each a key up or a key down event as it was earlier. It's makes no sense to cache the iterator, because the elements can get updated like when showing a file structure and enabling/disabling show inherited checkbox. The alarm now is used only for typing the search string.

GitOrigin-RevId: 77126b49f5180c18b4848fa18a45b2c91a64eb4a