[groovy] IDEA-272024: Don't show errors from GDSL files if they don't belong to the user's project
The issue is that EA-286433 has too many reports. We have no way to fix it, because the problem is on the user side in one of gdsl scripts.
Such a load of exceptions severely lowers the importance of the IDE internal error message, so we'll disable the possibility to get this exception if the user is not an owner of the culprit gdsl and cannot do anything about it themselves.

(cherry picked from commit 82656578fddc0557770465432b07c3b272b09c91)

IJ-CR-18303

GitOrigin-RevId: 57d5a8c6efd9c8c6efbadd252fec4091ccc354cf